# openapi.yml
openapi: 3.0.0
info:
  version: 0.1.0
  title: Social links service
  description: ''
paths:
  /api/v1/communications/send:
    post:
      summary: 'Отправить информацию о факте коммуникации'
      operationId: sendCommunication
      tags:
        - Communications
      requestBody:
        required: true
        description: 'Детали коммуникации. В текущей версии - только id участников (минимум двое)'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Communication'
      responses:
        '200':
          description: 'Данные приняты'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendCommunicationResponse'
        '422':
          description: 'Данные невалидные'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  '/api/v1/communications':
    get:
      summary: Получить гиперграф социальных связей
      operationId: getSocialGraph
      tags:
        - Communications
      responses:
        '200':
          description: 'Гиперграф успешно построен'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SocialGraph'
components:
  schemas:
    Communication:
      title: Communication
      type: object
      properties:
        user_ids:
          description: 'Участники коммуникации (минимум двое)'
          type: array
          items:
            type: integer
          minItems: 2
      required:
        - user_ids
    SocialGraph:
      title: SocialGraph
      description: 'Гиперграф социальных связей на основе коммуникаций'
      type: object
      required:
        - social_links
        - statistics
      properties:
        social_links:
          type: array
          items:
            $ref: '#/components/schemas/SocialLink'
        statistics:
          $ref: '#/components/schemas/Statistics'
    SocialLink:
      type: object
      title: SocialLink
      properties:
        user_ids:
          type: array
          items:
            type: integer
        communication_count:
          type: integer
      required:
        - user_ids
        - communication_count
    SendCommunicationResponse:
      title: SendCommunicationResponse
      type: object
      required:
        - communication_id
      properties:
        communication_id:
          type: integer
    Statistics:
      title: Statistics
      type: object
      required:
        - minimum_communication_count
        - average_communication_count
        - maximum_communication_count
      properties:
        minimum_communication_count:
          type: integer
        average_communication_count:
          type: number
        maximum_communication_count:
          type: integer
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            type: string
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
tags:
  - name: Communications