# src/social_links_service/codegen/models.py
# generated by fastapi-codegen:
#   filename:  openapi.yml
#   timestamp: 2022-09-13T15:02:06+00:00

from __future__ import annotations

from typing import List, Optional

from pydantic import BaseModel, Field


class Communication(BaseModel):
    user_ids: List[int] = Field(
        ..., description='Участники коммуникации (минимум двое)', min_items=2
    )


class SocialLink(BaseModel):
    user_ids: List[int]
    communication_count: int


class SendCommunicationResponse(BaseModel):
    communication_id: int


class Statistics(BaseModel):
    minimum_communication_count: int
    average_communication_count: float
    maximum_communication_count: int


class ValidationError(BaseModel):
    loc: List[str] = Field(..., title='Location')
    msg: str = Field(..., title='Message')
    type: str = Field(..., title='Error Type')


class SocialGraph(BaseModel):
    social_links: List[SocialLink]
    statistics: Statistics


class HTTPValidationError(BaseModel):
    detail: Optional[List[ValidationError]] = Field(None, title='Detail')